add docs, improve string fns
